+++ src/components/FlawForm.vue
@@ -22,7 +22,7 @@ import CvssCalculator from '@/components/CvssCalculator.vue';
 
 import { useFlawModel } from '@/composables/useFlawModel';
 import { fileTracker, type TrackersFilePost } from '@/services/TrackerService';
-import type { ZodFlawType } from '@/types/zodFlaw';
+import { type ZodFlawType, summaryRequiredStates } from '@/types/zodFlaw';
 
 const props = defineProps<{
   flaw: any;
@@ -108,6 +108,27 @@ const showMitigation = ref(flaw.value.mitigation && flaw.value.mitigation.trim()
 
 const onReset = () => {
   flaw.value = deepCopyFromRaw(initialFlaw.value as Record<string, any>) as ZodFlawType;
+};
+
+const toggleSummary = () => {
+  showSummary.value = !showSummary.value;
+  if (!showSummary.value) {
+    flaw.value.summary = '';
+  }
+};
+
+const toggleStatement = () => {
+  showStatement.value = !showStatement.value;
+  if (!showStatement.value) {
+    flaw.value.statement = '';
+  }
+};
+
+const toggleMitigation = () => {
+  showMitigation.value = !showMitigation.value;
+  if (!showMitigation.value) {
+    flaw.value.mitigation = '';
+  }
 };
 
 </script>
@@ -272,7 +293,21 @@ const onReset = () => {
           label="Description"
           placeholder="Description Text ..."
           :error="errors.summary"
-        />
+          class="osim-flaw-description-component"
+        >
+          <template #label>
+            <span class="form-label col-3 osim-folder-tab-label">
+              Description
+            </span>
+            <span class="col-3 ps-2">
+              <select v-model="flaw.requires_summary" class="form-select col-3 osim-summary-required">
+                <option disabled :selected="!flaw.requires_summary" value="">Review Status</option>
+                <option v-for="state in summaryRequiredStates" :key="state" :value="state">{{ state }}</option>
+              </select>
+            </span>
+
+          </template>
+        </LabelTextarea>
         <LabelTextarea
           v-if="showStatement"
           v-model="flaw.statement"
@@ -290,22 +325,22 @@ const onReset = () => {
         <div class="d-flex gap-3 mb-3">
           <button
             type="button"
-            class="btn btn-secondary"
-            @click="showSummary = !showSummary"
+            class="btn btn-secondary osim-show-summary"
+            @click="toggleSummary"
           >
             {{ showSummary ? 'Remove Description' : 'Add Description' }}
           </button>
           <button
             type="button"
             class="btn btn-secondary"
-            @click="showStatement = !showStatement"
+            @click="toggleStatement"
           >
             {{ showStatement ? 'Remove Statement' : 'Add Statement' }}
           </button>
           <button
             type="button"
             class="btn btn-secondary"
-            @click="showMitigation = !showMitigation"
+            @click="toggleMitigation"
           >
             {{ showMitigation ? 'Remove Mitigation' : 'Add Mitigation' }}
           </button>
@@ -452,7 +487,7 @@ form.osim-flaw-form :deep(*) {
     justify-content: end;
 
     &:has(+ textarea),
-    &:has(+ .osim-static-label-value.osim-top-label-style) {
+    &.osim-folder-tab-label {
       border-top-right-radius: 0.5rem;
       border-bottom-left-radius: 0;
       text-align: left;
@@ -462,9 +497,18 @@ form.osim-flaw-form :deep(*) {
   }
 
   textarea {
+    border-top-right-radius: 0.5rem;
     border-top-left-radius: 0;
   }
+
+  select.osim-summary-required.form-select {
+    border-bottom-right-radius: 0;
+    border-bottom-left-radius: 0;
+    margin-bottom: 0;
+    border-bottom: none;
+  }
 }
+
 
 .span-editable-text {
   cursor: text;